requirements.txt
aiofiles==23.2.1
aiogram==3.7.0
aiohttp==3.9.5
aiosignal==1.3.1
aiosqlite==0.20.0
alembic==1.14.0
annotated-types==0.7.0
async-timeout==4.0.3
asyncpg==0.29.0
attrs==23.2.0
black==24.4.2
certifi==2024.6.2
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
decouple==0.0.7
frozenlist==1.4.1
greenlet==3.0.3
idna==3.7
load-dotenv==0.1.0
loguru==0.7.2
magic-filter==1.0.12
Mako==1.3.6
MarkupSafe==3.0.2
multidict==6.0.5
mypy-extensions==1.0.0
packaging==24.1
pathspec==0.12.1
platformdirs==4.2.2
pydantic==2.7.4
pydantic_core==2.18.4
python-dotenv==1.0.1
requests==2.32.3
SQLAlchemy==2.0.31
typing_extensions==4.12.2
urllib3==2.2.2
win32-setctime==1.1.0
yarl==1.9.4
